requirements.txt
wheel==0.23.0
requests==2.7.0
twitter==1.17.0
facebook-sdk==0.4.0
